Ensure input array is chunked properly in linint2
If the input array is a Dask array but is chunked such that either of
the two rightmost dimensions are not entirely contiguous, rechunk the
array so that the chunked shape for those dimensions matches the full
length of those dimensions.

Also, removed fo.compute() call in linint2 so that Dask input arrays
will yield Dask output arrays. Calling the .compute() method results in
a numpy array being returned.