Added @carrayify decorator to Cython code
Using this decorator on a function in ncomp.pyx will ensure that all of
that function's numpy.ndarray arguments are C-contiguous in memory.

The decorator function iterates over each input argument, checks if the
argument is an instance of numpy.ndarray, then checks if its CARRAY flag
is set (meaning the C_CONTIGUOUS, ALIGNED, and BEHAVED flags are all
True), and then calls numpy.ascontiguousarray() on the argument array
and inserts the new contiguous array into the argument list in place of
the original non-contiguous array.

This decorator uses functools.wraps to ensure docstrings of all wrapped
functions are preserved.